Refactor app to follow app factory pattern
