+++ application/__init__.py
@@ -1,0 +1,16 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+def init_app():
+    app = Flask(__name__)
+    app.config.from_object('config.Config')
+    db.init_app(app)
+
+    with app.app_context():
+        from . import routes
+
+        db.create_all()
+
+        return app
