+++ wsgi.py
@@ -1,0 +1,6 @@
+from application import init_app
+
+app = init_app()
+
+if __name__ == "__main__":
+    app.run(debug=True)
